Implemented a SettingsHandler
It might be a bit over engineered, but... I just want simple access to my elements when coding, but a dynamic way to load them in.

Long story short, the `SettingsItem` stores the current value and a default value for if current value is None or empty (it _also_ might be empty but that's irrelevant). The `SettingsHandler` is an extension of a built-in `dict` enabling bracket set/retrieve, but that set/retrieve operates on the class instance attributes.

This allows for me/other devs to use dot notation + autocomplete, but also lets the loading and saving of the JSON to be simple and without excess repetition. I'm able to just loop through the keys and and restore/save the values to/from it very easily.

It's almost 130am, its functional, it's implemented, and the save/restore functionality works just fine. Further, this could help with a feature request I saw on the main github requesting profiles. Calling restore with a new JSON file to this, and then restore on all the elements... bam. Profile loaded. Would need to update the `Application.setting_file` and (potentially) the `setting_directory` but that's for when I implement it, not now.